+++ analysis_options.yaml
@@ -2,7 +2,6 @@ include: package:lint/analysis_options.yaml
 
 linter:
   rules:
-    file_names: false
     curly_braces_in_flow_control_structures: false
     avoid_classes_with_only_static_members: false
     sort_constructors_first: true
